+++ tests/chapter11/GuestbookIntegrationTest.php
@@ -1,0 +1,41 @@
+<?php
+/**
+ * Created by PhpStorm.
+ *
+ * PHP version 5.6, 7
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter11
+ * @author  Nikola Zeravcic <niks986@gmail.com>
+ * @license <http://opensource.org/licenses/gpl-license.php GPL
+ * @link    http://nikolazeravcic.iz.rs Personal site
+ */
+
+namespace Zeravcic\PhpUnit_De\Tests\chapter11;
+
+use PHPUnit\Framework\TestCase;
+
+/**
+ * Class GuestbookIntegrationTest
+ *
+ * @package Zeravcic\PhpUnit_De\Tests\chapter11
+ */
+class GuestbookIntegrationTest extends PHPUnit_Extensions_Database_TestCase
+{
+    /**
+     * @coversNothing
+     */
+    public function testAddEntry()
+    {
+        $guestbook = new Guestbook();
+        $guestbook->addEntry("suzy", "Hello world!");
+
+        $queryTable = $this->getConnection()->createQueryTable(
+            'guestbook', 'SELECT * FROM guestbook'
+        );
+
+        $expectedTable = $this->createFlatXmlDataSet("expectedBook.xml")
+            ->getTable("guestbook");
+
+        $this->assertTablesEqual($expectedTable, $queryTable);
+    }
+}